feat: add advanced signature editor UI and split views
- Add new editor page at /app/editor/page.tsx with full signature
  building functionality
- Split single page into landing (quick mode) and advanced editor views
- Update routing to support ui=editor parameter in short share links
- Redirect /{text} URLs to appropriate UI based on ui parameter
- Enhance GIF generation with proper animated implementation using
  timeline sampling
- Add mobile-responsive layout with resizable panels and drawer sidebar
- Implement GIF export dialog with FPS and quality controls
- Update README documentation to reflect new architecture and data flows
- Add .ace-tool to gitignore
