+++ .gitignore
@@ -63,3 +63,4 @@ next-env.d.ts
 
 .claude/
 CLAUDE.md
+.ace-tool/
